Support multiple urls per backend

+++ backend_storage_static.go
@@ -24,7 +24,7 @@ package signaling
 import (
 	"log"
 	"net/url"
-	"reflect"
+	"slices"
 	"strings"
 
 	"github.com/dlintw/goconf"
@@ -32,6 +32,8 @@ import (
 
 type backendStorageStatic struct {
 	backendStorageCommon
+
+	backendsById map[string]*Backend
 
 	// Deprecated
 	allowAll      bool
@@ -48,6 +50,7 @@ func NewBackendStorageStatic(config *goconf.ConfigFile) (BackendStorage, error) 
 		sessionLimit = 0
 	}
 	backends := make(map[string][]*Backend)
+	backendsById := make(map[string]*Backend)
 	var compatBackend *Backend
 	numBackends := 0
 	if allowAll {
@@ -55,26 +58,33 @@ func NewBackendStorageStatic(config *goconf.ConfigFile) (BackendStorage, error) 
 		compatBackend = &Backend{
 			id:     "compat",
 			secret: []byte(commonSecret),
-			compat: true,
 
 			allowHttp: allowHttp,
 
 			sessionLimit: uint64(sessionLimit),
+			counted:      true,
 		}
 		if sessionLimit > 0 {
 			log.Printf("Allow a maximum of %d sessions", sessionLimit)
 		}
 		updateBackendStats(compatBackend)
+		backendsById[compatBackend.id] = compatBackend
 		numBackends++
 	} else if backendIds, _ := config.GetString("backend", "backends"); backendIds != "" {
+		added := make(map[string]*Backend)
 		for host, configuredBackends := range getConfiguredHosts(backendIds, config, commonSecret) {
 			backends[host] = append(backends[host], configuredBackends...)
 			for _, be := range configuredBackends {
-				log.Printf("Backend %s added for %s", be.id, be.url)
-				updateBackendStats(be)
+				added[be.id] = be
 			}
-			numBackends += len(configuredBackends)
 		}
+		for _, be := range added {
+			log.Printf("Backend %s added for %s", be.id, strings.Join(be.urls, ", "))
+			backendsById[be.id] = be
+			updateBackendStats(be)
+			be.counted = true
+		}
+		numBackends += len(added)
 	} else if allowedUrls, _ := config.GetString("backend", "allowed"); allowedUrls != "" {
 		// Old-style configuration, only hosts are configured and are using a common secret.
 		allowMap := make(map[string]bool)
@@ -95,11 +105,11 @@ func NewBackendStorageStatic(config *goconf.ConfigFile) (BackendStorage, error) 
 			compatBackend = &Backend{
 				id:     "compat",
 				secret: []byte(commonSecret),
-				compat: true,
 
 				allowHttp: allowHttp,
 
 				sessionLimit: uint64(sessionLimit),
+				counted:      true,
 			}
 			hosts := make([]string, 0, len(allowMap))
 			for host := range allowMap {
@@ -114,6 +124,7 @@ func NewBackendStorageStatic(config *goconf.ConfigFile) (BackendStorage, error) 
 				log.Printf("Allow a maximum of %d sessions", sessionLimit)
 			}
 			updateBackendStats(compatBackend)
+			backendsById[compatBackend.id] = compatBackend
 			numBackends++
 		}
 	}
@@ -128,6 +139,8 @@ func NewBackendStorageStatic(config *goconf.ConfigFile) (BackendStorage, error) 
 			backends: backends,
 		},
 
+		backendsById: backendsById,
+
 		allowAll:      allowAll,
 		commonSecret:  []byte(commonSecret),
 		compatBackend: compatBackend,
@@ -137,23 +150,56 @@ func NewBackendStorageStatic(config *goconf.ConfigFile) (BackendStorage, error) 
 func (s *backendStorageStatic) Close() {
 }
 
-func (s *backendStorageStatic) RemoveBackendsForHost(host string) {
+func (s *backendStorageStatic) RemoveBackendsForHost(host string, seen map[string]seenState) {
 	if oldBackends := s.backends[host]; len(oldBackends) > 0 {
+		deleted := 0
 		for _, backend := range oldBackends {
-			log.Printf("Backend %s removed for %s", backend.id, backend.url)
-			deleteBackendStats(backend)
+			if seen[backend.Id()] == seenDeleted {
+				continue
+			}
+
+			seen[backend.Id()] = seenDeleted
+			urls := filter(backend.urls, func(s string) bool {
+				return !strings.Contains(s, "://"+host)
+			})
+			log.Printf("Backend %s removed for %s", backend.id, strings.Join(urls, ", "))
+			if len(urls) == len(backend.urls) && backend.counted {
+				deleteBackendStats(backend)
+				delete(s.backendsById, backend.Id())
+				deleted++
+				backend.counted = false
+			}
 		}
-		statsBackendsCurrent.Sub(float64(len(oldBackends)))
+		statsBackendsCurrent.Sub(float64(deleted))
 	}
 	delete(s.backends, host)
 }
 
-func (s *backendStorageStatic) UpsertHost(host string, backends []*Backend) {
+func filter[T any](s []T, del func(T) bool) []T {
+	result := make([]T, 0, len(s))
+	for _, e := range s {
+		if !del(e) {
+			result = append(result, e)
+		}
+	}
+	return result
+}
+
+type seenState int
+
+const (
+	seenNotSeen seenState = iota
+	seenAdded
+	seenUpdated
+	seenDeleted
+)
+
+func (s *backendStorageStatic) UpsertHost(host string, backends []*Backend, seen map[string]seenState) {
 	for existingIndex, existingBackend := range s.backends[host] {
 		found := false
 		index := 0
 		for _, newBackend := range backends {
-			if reflect.DeepEqual(existingBackend, newBackend) { // otherwise we could manually compare the struct members here
+			if existingBackend.Equal(newBackend) {
 				found = true
 				backends = append(backends[:index], backends[index+1:]...)
 				break
@@ -161,27 +207,59 @@ func (s *backendStorageStatic) UpsertHost(host string, backends []*Backend) {
 				found = true
 				s.backends[host][existingIndex] = newBackend
 				backends = append(backends[:index], backends[index+1:]...)
-				log.Printf("Backend %s updated for %s", newBackend.id, newBackend.url)
-				updateBackendStats(newBackend)
+				if seen[newBackend.id] != seenUpdated {
+					seen[newBackend.id] = seenUpdated
+					log.Printf("Backend %s updated for %s", newBackend.id, strings.Join(newBackend.urls, ", "))
+					updateBackendStats(newBackend)
+					newBackend.counted = existingBackend.counted
+					s.backendsById[newBackend.id] = newBackend
+				}
 				break
 			}
 			index++
 		}
 		if !found {
 			removed := s.backends[host][existingIndex]
-			log.Printf("Backend %s removed for %s", removed.id, removed.url)
 			s.backends[host] = append(s.backends[host][:existingIndex], s.backends[host][existingIndex+1:]...)
-			deleteBackendStats(removed)
-			statsBackendsCurrent.Dec()
+			if seen[removed.id] != seenDeleted {
+				seen[removed.id] = seenDeleted
+				urls := filter(removed.urls, func(s string) bool {
+					return !strings.Contains(s, "://"+host)
+				})
+				log.Printf("Backend %s removed for %s", removed.id, strings.Join(urls, ", "))
+				if len(urls) == len(removed.urls) && removed.counted {
+					deleteBackendStats(removed)
+					delete(s.backendsById, removed.Id())
+					statsBackendsCurrent.Dec()
+					removed.counted = false
+				}
+			}
 		}
 	}
 
 	s.backends[host] = append(s.backends[host], backends...)
+
+	addedBackends := 0
 	for _, added := range backends {
-		log.Printf("Backend %s added for %s", added.id, added.url)
-		updateBackendStats(added)
+		if seen[added.id] == seenAdded {
+			continue
+		}
+
+		seen[added.id] = seenAdded
+		if prev, found := s.backendsById[added.id]; found {
+			added.counted = prev.counted
+		} else {
+			s.backendsById[added.id] = added
+		}
+
+		log.Printf("Backend %s added for %s", added.id, strings.Join(added.urls, ", "))
+		if !added.counted {
+			updateBackendStats(added)
+			addedBackends++
+			added.counted = true
+		}
 	}
-	statsBackendsCurrent.Add(float64(len(backends)))
+	statsBackendsCurrent.Add(float64(addedBackends))
 }
 
 func getConfiguredBackendIDs(backendIds string) (ids []string) {
@@ -203,35 +281,26 @@ func getConfiguredBackendIDs(backendIds string) (ids []string) {
 	return ids
 }
 
+func MapIf[T any](s []T, f func(T) (T, bool)) []T {
+	result := make([]T, 0, len(s))
+	for _, v := range s {
+		if v, ok := f(v); ok {
+			result = append(result, v)
+		}
+	}
+	return result
+}
+
 func getConfiguredHosts(backendIds string, config *goconf.ConfigFile, commonSecret string) (hosts map[string][]*Backend) {
 	hosts = make(map[string][]*Backend)
+	seenUrls := make(map[string]string)
 	for _, id := range getConfiguredBackendIDs(backendIds) {
-		u, _ := GetStringOptionWithEnv(config, id, "url")
-		if u == "" {
-			log.Printf("Backend %s is missing or incomplete, skipping", id)
-			continue
-		}
-
-		if u[len(u)-1] != '/' {
-			u += "/"
-		}
-		parsed, err := url.Parse(u)
-		if err != nil {
-			log.Printf("Backend %s has an invalid url %s configured (%s), skipping", id, u, err)
-			continue
-		}
-
-		if strings.Contains(parsed.Host, ":") && hasStandardPort(parsed) {
-			parsed.Host = parsed.Hostname()
-			u = parsed.String()
-		}
-
 		secret, _ := GetStringOptionWithEnv(config, id, "secret")
 		if secret == "" && commonSecret != "" {
 			log.Printf("Backend %s has no own shared secret set, using common shared secret", id)
 			secret = commonSecret
 		}
-		if u == "" || secret == "" {
+		if secret == "" {
 			log.Printf("Backend %s is missing or incomplete, skipping", id)
 			continue
 		}
@@ -253,19 +322,69 @@ func getConfiguredHosts(backendIds string, config *goconf.ConfigFile, commonSecr
 			maxScreenBitrate = 0
 		}
 
-		hosts[parsed.Host] = append(hosts[parsed.Host], &Backend{
-			id:        id,
-			url:       u,
-			parsedUrl: parsed,
-			secret:    []byte(secret),
+		var urls []string
+		if u, _ := GetStringOptionWithEnv(config, id, "urls"); u != "" {
+			urls = strings.Split(u, ",")
+			urls = MapIf(urls, func(s string) (string, bool) {
+				s = strings.TrimSpace(s)
+				return s, len(s) > 0
+			})
+			slices.Sort(urls)
+			urls = slices.Compact(urls)
+		} else if u, _ := GetStringOptionWithEnv(config, id, "url"); u != "" {
+			if u = strings.TrimSpace(u); u != "" {
+				urls = []string{u}
+			}
+		}
 
-			allowHttp: parsed.Scheme == "http",
+		if len(urls) == 0 {
+			log.Printf("Backend %s is missing or incomplete, skipping", id)
+			continue
+		}
+
+		backend := &Backend{
+			id:     id,
+			secret: []byte(secret),
 
 			maxStreamBitrate: maxStreamBitrate,
 			maxScreenBitrate: maxScreenBitrate,
 
 			sessionLimit: uint64(sessionLimit),
-		})
+		}
+
+		added := make(map[string]bool)
+		for _, u := range urls {
+			if u[len(u)-1] != '/' {
+				u += "/"
+			}
+
+			parsed, err := url.Parse(u)
+			if err != nil {
+				log.Printf("Backend %s has an invalid url %s configured (%s), skipping", id, u, err)
+				continue
+			}
+
+			if strings.Contains(parsed.Host, ":") && hasStandardPort(parsed) {
+				parsed.Host = parsed.Hostname()
+				u = parsed.String()
+			}
+
+			if prev, found := seenUrls[u]; found {
+				log.Printf("Url %s in backend %s was already used in backend %s, skipping", u, id, prev)
+				continue
+			}
+
+			seenUrls[u] = id
+			backend.urls = append(backend.urls, u)
+			if parsed.Scheme == "http" {
+				backend.allowHttp = true
+			}
+
+			if !added[parsed.Host] {
+				hosts[parsed.Host] = append(hosts[parsed.Host], backend)
+				added[parsed.Host] = true
+			}
+		}
 	}
 
 	return hosts
@@ -286,15 +405,22 @@ func (s *backendStorageStatic) Reload(config *goconf.ConfigFile) {
 		configuredHosts := getConfiguredHosts(backendIds, config, commonSecret)
 
 		// remove backends that are no longer configured
+		seen := make(map[string]seenState)
 		for hostname := range s.backends {
 			if _, ok := configuredHosts[hostname]; !ok {
-				s.RemoveBackendsForHost(hostname)
+				s.RemoveBackendsForHost(hostname, seen)
 			}
 		}
 
 		// rewrite backends adding newly configured ones and rewriting existing ones
 		for hostname, configuredBackends := range configuredHosts {
-			s.UpsertHost(hostname, configuredBackends)
+			s.UpsertHost(hostname, configuredBackends, seen)
+		}
+	} else {
+		// remove all backends
+		seen := make(map[string]seenState)
+		for hostname := range s.backends {
+			s.RemoveBackendsForHost(hostname, seen)
 		}
 	}
 }
